extract ExtraMeta to a separate file
Summary:

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:
