+++ c10/core/TensorImpl.h
@@ -10,7 +10,9 @@
 #include <c10/core/SymIntArrayRef.h>
 #include <c10/core/TensorOptions.h>
 #include <c10/core/WrapDimMinimal.h>
+#include <c10/core/impl/ExtraMeta.h>
 #include <c10/core/impl/LocalDispatchKeySet.h>
+#include <c10/core/impl/NamedTensorMetaInterface.h>
 #include <c10/core/impl/PyObjectSlot.h>
 #include <c10/core/impl/SizesAndStrides.h>
 #include <c10/util/DimVector.h>
@@ -195,18 +197,6 @@ struct C10_API AutogradMetaFactoryRegisterer {
 
 } // namespace impl
 
-struct C10_API NamedTensorMetaInterface {
-  virtual ~NamedTensorMetaInterface() = default;
-  virtual std::unique_ptr<NamedTensorMetaInterface> clone() const {
-    TORCH_INTERNAL_ASSERT(
-        false, "Not implemented: NamedTensorMetaInterface::clone");
-  };
-  virtual int64_t slow_dim() const {
-    TORCH_INTERNAL_ASSERT(
-        false, "Not implemented: NamedTensorMetaInterface::slow_dim");
-  };
-};
-
 // For ease of copy pasting
 #if 0
 is_contiguous
@@ -216,62 +206,6 @@ is_channels_last
 is_channels_last_3d
 is_non_overlapping_and_dense
 #endif
-
-struct C10_API ExtraMeta {
-  SymDimVector sizes_ = {0};
-  SymDimVector strides_ = {1};
-  SymInt numel_ = 1;
-  SymInt storage_offset_ = 0;
-  SymBool is_contiguous_{true};
-  SymBool is_channels_last_contiguous_{false};
-  SymBool is_channels_last_3d_contiguous_{false};
-  SymBool is_channels_last_{false};
-  SymBool is_channels_last_3d_{false};
-  SymBool is_non_overlapping_and_dense_{true};
-  std::unique_ptr<c10::NamedTensorMetaInterface> named_tensor_meta_ = nullptr;
-
-  ExtraMeta() = default;
-
-  ExtraMeta(
-      SymDimVector sizes,
-      SymDimVector strides,
-      SymInt numel,
-      SymInt storage_offset,
-      SymBool is_contiguous,
-      SymBool is_channels_last_contiguous,
-      SymBool is_channels_last_3d_contiguous,
-      SymBool is_channels_last,
-      SymBool is_channels_last_3d,
-      SymBool is_non_overlapping_and_dense,
-      std::unique_ptr<c10::NamedTensorMetaInterface> named_tensor_meta)
-      : sizes_(std::move(sizes)),
-        strides_(std::move(strides)),
-        numel_(std::move(numel)),
-        storage_offset_(std::move(storage_offset)),
-        is_contiguous_(std::move(is_contiguous)),
-        is_channels_last_contiguous_(std::move(is_channels_last_contiguous)),
-        is_channels_last_3d_contiguous_(
-            std::move(is_channels_last_3d_contiguous)),
-        is_channels_last_(std::move(is_channels_last)),
-        is_channels_last_3d_(std::move(is_channels_last_3d)),
-        is_non_overlapping_and_dense_(std::move(is_non_overlapping_and_dense)),
-        named_tensor_meta_(std::move(named_tensor_meta)) {}
-
-  std::unique_ptr<ExtraMeta> clone() const {
-    return std::make_unique<ExtraMeta>(
-        sizes_,
-        strides_,
-        numel_,
-        storage_offset_,
-        is_contiguous_,
-        is_channels_last_contiguous_,
-        is_channels_last_3d_contiguous_,
-        is_channels_last_,
-        is_channels_last_3d_,
-        is_non_overlapping_and_dense_,
-        named_tensor_meta_ ? named_tensor_meta_->clone() : nullptr);
-  }
-};
 
 // NOTE [ Version Counter Sharing ]
 //
@@ -1796,7 +1730,7 @@ struct C10_API TensorImpl : public c10::intrusive_ptr_target {
 #endif
     if (named_tensor_meta) {
       if (!extra_meta_) {
-        extra_meta_ = std::make_unique<ExtraMeta>();
+        extra_meta_ = std::make_unique<impl::ExtraMeta>();
       }
       extra_meta_->named_tensor_meta_ = std::move(named_tensor_meta);
       key_set_ = key_set_.add(DispatchKey::Named);
@@ -2775,7 +2709,7 @@ struct C10_API TensorImpl : public c10::intrusive_ptr_target {
   std::unique_ptr<c10::AutogradMetaInterface> autograd_meta_ = nullptr;
 
  protected:
-  std::unique_ptr<c10::ExtraMeta> extra_meta_ = nullptr;
+  std::unique_ptr<impl::ExtraMeta> extra_meta_ = nullptr;
 
   c10::VariableVersion version_counter_;
